+++ raydium/light.c
@@ -87,7 +87,7 @@ raydium_light_update_position(l);
 
 signed char raydium_light_texture(int texture, signed char enable)
 {
-if(texture>=0 && texture<raydium_texture_index)
+if(texture>=0 && texture<(int)raydium_texture_index)
     {
     raydium_texture_nolight[texture]=!enable;
     return 1;
